chore: todoList -> roleNrule 폴더 네이밍 변경
