+++ src/components/roleNrule/selectMate.tsx
@@ -1,50 +1,39 @@
 import React from 'react';
 import { Text, View, Pressable } from 'react-native';
 
-import Selected from '@assets/todoList/selectedCheckBox.svg';
-import NotSelected from '@assets/todoList/notSelectedCheckBox.svg';
-
-interface RoleMateItem {
-  mateId: number;
-  nickname: string;
-}
+import Selected from '@assets/roleNrule/selectedCheckBox.svg';
+import NotSelected from '@assets/roleNrule/notSelectedCheckBox.svg';
 
 interface SelectMateComponentProps {
   title: string;
-  selectedValues: RoleMateItem[];
-  setSelectedValues: React.Dispatch<React.SetStateAction<RoleMateItem[]>>;
+  selectedValues: number[];
+  setSelectedValues: React.Dispatch<React.SetStateAction<number[]>>;
   items: { memberId: number; mateId: number; nickname: string }[];
 }
 
-const RoleSelectMateComponent: React.FC<SelectMateComponentProps> = ({
+const SelectMateComponent: React.FC<SelectMateComponentProps> = ({
   title,
   selectedValues,
   setSelectedValues,
   items,
 }) => {
-  const toggleSelection = (item: { mateId: number; nickname: string }) => {
-    const isSelected = selectedValues.some((value) => value.mateId === item.mateId);
+  const toggleSelection = (item: { mateId: number }) => {
+    const itemId = item.mateId;
+    const isSelected = selectedValues.includes(itemId);
 
     const updatedSelectedValues = isSelected
-      ? selectedValues.filter((value) => value.mateId !== item.mateId)
-      : [...selectedValues, { mateId: item.mateId, nickname: item.nickname }];
+      ? selectedValues.filter((value) => value !== itemId)
+      : [...selectedValues, itemId];
 
     setSelectedValues(updatedSelectedValues);
   };
 
   const toggleAllItems = () => {
-    const areAllSelected = items.every((item) =>
-      selectedValues.some((value) => value.mateId === item.mateId),
-    );
-
-    setSelectedValues(
-      areAllSelected ? [] : items.map((item) => ({ mateId: item.mateId, nickname: item.nickname })),
-    );
+    const areAllSelected = items.every((item) => selectedValues.includes(item.mateId));
+    setSelectedValues(areAllSelected ? [] : items.map((item) => item.mateId));
   };
 
-  const isEveryItemSelected = items.every((item) =>
-    selectedValues.some((value) => value.mateId === item.mateId),
-  );
+  const isEveryItemSelected = items.every((item) => selectedValues.includes(item.mateId));
 
   return (
     <View className="mb-10">
@@ -54,15 +43,13 @@ const RoleSelectMateComponent: React.FC<SelectMateComponentProps> = ({
           <Pressable
             key={item.mateId}
             className={`mb-2 flex items-center justify-center rounded-md px-5 py-[10px] ${
-              selectedValues.some((value) => value.mateId === item.mateId)
-                ? 'bg-sub1'
-                : 'bg-colorBox'
-            }`}
+              selectedValues.includes(item.mateId) ? 'bg-sub1' : 'bg-colorBox'
+            } `}
             onPress={() => toggleSelection(item)}
           >
             <Text
               className={`text-sm tracking-tight ${
-                selectedValues.some((value) => value.mateId === item.mateId)
+                selectedValues.includes(item.mateId)
                   ? 'font-semibold text-main1'
                   : 'font-medium text-disabledFont'
               }`}
@@ -88,4 +75,4 @@ const RoleSelectMateComponent: React.FC<SelectMateComponentProps> = ({
   );
 };
 
-export default RoleSelectMateComponent;
+export default SelectMateComponent;
